[ImageUtil] Code clean (#4887)
